comment post like 셋의 연관관계 수정 application.yml ddl-auto: create-drop 상태 주의

+++ src/main/kotlin/com/sparta/techTree/comment/dto/CommentResponse.kt
@@ -9,5 +9,5 @@ data class CommentResponse(
     val content: String,
     val createdAt: LocalDateTime,
     val updatedAt: LocalDateTime,
-    val countLikes: Long,
+    val countLikes: Int
 )
